+++ R/utils.R
@@ -32,13 +32,13 @@ create_index <- function(x, path = ".") {
 
   index <- htmltools::htmlTemplate(
     system.file(paste0(x, "/index.html"), package = "altdoc"),
-    title = pkg_name(),
+    title = pkg_name(path),
     footer = sprintf(
       "<hr/><a href='%s'> <code>%s</code> v. %s </a> | Documentation made with <a href='https://github.com/etiennebacher/altdoc'> <code>altdoc</code> v. %s</a>",
-      gh_url(), pkg_name(), pkg_version(),
-      utils::packageVersion("altdoc")
+      gh_url(path), pkg_name(path), pkg_version(path),
+      altdoc_version()
     ),
-    github_link = gh_url()
+    github_link = gh_url(path)
   )
 
   # regex stuff to correct footer
@@ -264,34 +264,28 @@ stop_quietly <- function() {
 # @param x Name of the folder
 
 folder_is_empty <- function(x) {
-
-  if (length(list.files(x)) == 0) {
-    return(TRUE)
-  } else {
-    return(FALSE)
-  }
-
+  length(list.files(x)) == 0
 }
 
 # Get package name
-pkg_name <- function() {
-  desc::desc_get_field("Package", default = NULL)
+pkg_name <- function(path) {
+  desc::desc_get_field("Package", default = NULL, file = path)
 }
 
 # Get package version
-pkg_version <- function() {
-  as.character(desc::desc_get_version())
+pkg_version <- function(path) {
+  as.character(desc::desc_get_version(path))
 }
 
 # Get package Github URL
-gh_url <- function() {
+gh_url <- function(path) {
 
   gh_urls <- c(
-    desc::desc_get_urls(),
-    desc::desc_get_field("BugReports", default = NULL)
+    desc::desc_get_urls(path),
+    desc::desc_get_field("BugReports", default = NULL, file = path)
   )
 
-  if (length(gh_urls) == 0) return(NULL)
+  if (length(gh_urls) == 0) return("")
 
   gh_url <- gh_urls[which(grepl("github.com", gh_urls))]
   gh_url <- gsub("/issues", "", gh_url)
@@ -302,8 +296,10 @@ gh_url <- function() {
   }
   gh_url <- gsub(" ", "", gh_url)
   gh_url <- gsub("#.*", "", gh_url)
+  gh_url <- unique(gh_url)
 
-  return(unique(gh_url))
+  if (length(gh_url) == 0) gh_url <- ""
+  return(gh_url)
 
 }
 
@@ -383,6 +379,49 @@ check_is_package <- function(path) {
     cli::cli_alert_danger("{.code altdoc} only works in packages.")
     stop_quietly()
   }
+}
+
+get_footer <- function(path) {
+  doc_type <- doc_type(path)
+  if (doc_type %in% c("docute", "docsify")) {
+    index <- .readlines("docs/index.html")
+    index <- gsub("\\t", "", index)
+    index <- trimws(index)
+    if (doc_type == "docsify") {
+      footer <- which(grepl("^var footer =", index))
+    } else if (doc_type == "docute") {
+      footer <- which(grepl("^footer:", index))
+    }
+    if (length(footer) == 1) {
+      return(index[footer])
+    }
+  }
+}
+
+doc_version <- function(path) {
+  footer <- get_footer(path)
+  unlist(regmatches(
+    footer, gregexpr("(\\d+\\.\\d+\\.\\d+(?:\\.\\d+)?)", footer)
+  ))[1]
+}
+
+altdoc_version_in_footer <- function(path) {
+  footer <- get_footer(path)
+  unlist(regmatches(
+    footer, gregexpr("(\\d+\\.\\d+\\.\\d+(?:\\.\\d+)?)", footer)
+  ))[2]
+}
+
+altdoc_version <- function() {
+  as.character(utils::packageVersion("altdoc"))
+}
+
+need_to_bump_version <- function(path) {
+  doc_version(path) != pkg_version(path)
+}
+
+need_to_bump_altdoc_version <- function(path) {
+  altdoc_version_in_footer(path) != altdoc_version()
 }
 
 .readlines <- function(x) {