+++ docs/NEWS.md
@@ -1,3 +1,9 @@
-# altdoc 0.1.0
+## altdoc 0.1.0.1 (development)
 
-* First version.
+- `update_docs()` now updates the package version as well as altdoc version in the footer.
+
+## altdoc 0.1.0
+
+- First version.
+
+